requirements.txt
streamlit
langchain
langchain_core
langchain_community
langchain_openai
langchain-groq
langchain-nvidia-ai-endpoints

groq
python-dotenv
pdf2image
google-generativeai

numpy
Pillow
Requests
transformers
accelerate
streamlit 
torch 
torchvision 
transformers 
sentence-transformers 
faiss-cpu
pytesseract
vertexai